make some updates
