+++ lib/features/auth_by_email/view/sign_up.dart
@@ -1,15 +1,16 @@
 import 'package:flutter/material.dart';
+import 'package:galaxyxgate/core/themes/app_colors.dart';
 import 'package:galaxyxgate/core/widgets/continue_button.dart';
-import 'package:galaxyxgate/features/auth_by_email/screen/sign_in_screen.dart';
-import 'package:galaxyxgate/features/auth_by_email/screen/widgets/sign_up_text_and_fields.dart';
+import 'package:galaxyxgate/features/auth_by_email/view/sign_in.dart';
+import 'package:galaxyxgate/features/auth_by_email/view/widgets/sign_up_texts_and_fields.dart';
 
-class SignUpScreen extends StatelessWidget {
-  const SignUpScreen({super.key});
+class SignUp extends StatelessWidget {
+  const SignUp({super.key});
   @override
   Widget build(BuildContext context) {
     double screenHeight = MediaQuery.of(context).size.height;
     return Scaffold(
-      backgroundColor: Colors.black,
+      backgroundColor: AppColors.black,
       body: SafeArea(
         child: SingleChildScrollView(
           child: Column(
@@ -24,7 +25,7 @@ class SignUpScreen extends StatelessWidget {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                        builder: (context) => const SignInScreen(),
+                        builder: (context) => const SignIn(),
                       ),
                    );
                   },
