+++ lib/features/auth_by_email/view/widgets/custom_text_field.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:galaxyxgate/core/themes/app_colors.dart';
-
+import 'package:galaxyxgate/core/themes/text_styles.dart';
 
 class CustomTextField extends StatelessWidget {
   const CustomTextField(
@@ -10,16 +10,12 @@ class CustomTextField extends StatelessWidget {
   final Icon? icon;
   @override
   Widget build(BuildContext context) {
-  double screenHeight = MediaQuery.of(context).size.height;
+    double screenHeight = MediaQuery.of(context).size.height;
     return SizedBox(
       width: screenHeight * 0.43,
       child: TextField(
-        style: const TextStyle(
-          color: Colors.white,
-          fontSize: 14,
-          height: 3,
-        ),
-        cursorColor: Colors.white,
+        style: TextStyles.textStyle14400.copyWith(height: 2),
+        cursorColor: AppColors.white,
         decoration: InputDecoration(
           focusedBorder: OutlineInputBorder(
             borderRadius: BorderRadius.circular(15),
@@ -36,24 +32,12 @@ class CustomTextField extends StatelessWidget {
             ),
           ),
           filled: true,
-          fillColor: const Color(0xff3d3d3d33),
+          fillColor: AppColors.black,
           hintText: hintText,
           labelText: labelText,
-          hintStyle: const TextStyle(
-            fontSize: 12,
-            fontWeight: FontWeight.w400,
-            color: Color(0xff686868),
-            fontFamily: 'SpaceGrotesk',
-          ),
-          labelStyle: const TextStyle(
-            fontSize: 14,
-            fontWeight: FontWeight.w400,
-            color: Color(0xffFFFDF6),
-            fontFamily: 'SpaceGrotesk',
-          ),
-          
+          hintStyle:  TextStyles.textStyle12400,
+          labelStyle:  TextStyles.textStyle14400.copyWith(color:AppColors.white),
         ),
-        
       ),
     );
   }
